pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "numpy>=2.4.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
